requirements.txt
tencentcloud-sdk-python
alibabacloud_ocr_api20210707==3.1.2
alibabacloud-tea-openapi
alibabacloud-tea-util
PyMuPDF
Pillow
pytest
openpyxl
PyYAML
urllib3<2
